+++ api/models/round.go
@@ -1,8 +1,8 @@
 package models
 
 type Round struct {
-	ID     int64  `json:"ID"`
-	UID    int64  `json:"UID"`
+	ID     int    `json:"ID"`
+	UID    int    `json:"UID"`
 	Course string `json:"Course"`
-	Score  int16  `json:"Score"`
+	Score  int    `json:"Score"`
 }
